add component denymodal
