+++ src/components/DenyApplicationModal/DenyApplicationModal.module.css
@@ -1,0 +1,73 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 80vh;
+}
+
+.modalTitle {
+  font-weight: 700;
+  font-size: 16px;
+  line-height: 147%;
+  color: #1c1c1e;
+}
+
+.modalText {
+  font-family: Ubuntu;
+  font-weight: 500;
+  font-size: 16px;
+  line-height: 147%;
+  color: #1a202c;
+  margin: 16px 0 32px 0;
+}
+
+.modalButtons {
+  display: flex;
+  justify-content: right;
+  gap: 32px;
+  width: 100%;
+}
+
+.modalCancelButton,
+.modalGoHomeButton {
+  padding: 12px 24px;
+  border-radius: 6px;
+  font-size: 16px;
+  font-weight: 700;
+  cursor: pointer;
+  transition: background-color 0.2s;
+  border: none;
+  color: #fff;
+}
+
+.modalCancelButton,
+.modalGoHomeButton {
+  background-color: #003cff;
+  color: white;
+  border: none;
+  padding: 12px 24px;
+  border-radius: 8px;
+  cursor: pointer;
+  font-weight: 700;
+  font-size: 16px;
+}
+.modalGoHomeButton {
+  float: right;
+}
+
+.document__denyButton,
+.modalDenyButton {
+  background-color: #d93737cc;
+  color: white;
+  border: none;
+  padding: 12px 24px;
+  border-radius: 8px;
+  cursor: pointer;
+  font-weight: 700;
+  font-size: 16px;
+}
+
+button:hover {
+  filter: brightness(1.4);
+}
